Fix: Styles of previous version with new structure
