+++ src/app/page.tsx
@@ -1,51 +1,17 @@
-"use client"
-import { GameCard } from "~/components/GameCard"
 import { LoadingSpinner } from "~/components/ui-blocks/LoadingSpinner"
-import { Pagination } from "~/components/ui-blocks/Pagination"
-import { useGamesQuery } from "~/graphql/generated"
-import { usePage } from "~/hooks/usePage"
-import { useSearchParams } from "next/navigation"
+import { Suspense } from "react"
+import { GamesListPageContent } from "~/components/page-content/GamesListPageContent"
 
 export default function GamesListPage() {
-  const searchParams = useSearchParams()
-  const page = usePage(searchParams)
-
-  const { data, loading, error } = useGamesQuery({
-    variables: {
-      first: page.itemsPerPage,
-      skip: page.itemsToSkip,
-    },
-  })
-
-  if (loading) {
-    return (
-      <div className="grid h-screen place-items-center">
-        <LoadingSpinner />
-      </div>
-    )
-  }
-
-  if (error) {
-    return <h2>Algo salió mal</h2>
-  }
-
   return (
-    <div className="flex flex-col items-center justify-center p-4 sm:p-5 lg:px-20">
-      <div className="my-5 flex flex-col gap-2 text-center">
-        <h1 className="text-2xl font-semibold text-gray-900 lg:text-3xl">Lista de juegos</h1>
-        <h4 className="italic">Página {page.page}</h4>
-      </div>
-
-      {data && (
-        <section className="flex flex-col items-center">
-          <div className="grid grid-cols-1 gap-5 md:grid-cols-2 lg:grid-cols-3">
-            {data.games.map((game) => {
-              return <GameCard key={game.id} game={game} />
-            })}
-          </div>
-          <Pagination page={page} pageMetaData={data.gamesConnection} pathname="/" />
-        </section>
-      )}
-    </div>
+    <Suspense
+      fallback={
+        <div className="grid h-screen place-items-center">
+          <LoadingSpinner />
+        </div>
+      }
+    >
+      <GamesListPageContent />
+    </Suspense>
   )
 }
