+++ src/components/ui-blocks/Layout/index.tsx
@@ -8,9 +8,9 @@ type LayoutProps = {
 
 export function Layout({ children }: LayoutProps) {
   return (
-    <div className="flex h-screen flex-col justify-between">
+    <div className="min-h-screen flex flex-col">
       <Header />
-      <main className="mb-auto">{children}</main>
+      <main className="flex-1 w-full">{children}</main>
       <Footer />
     </div>
   )
