+++ tailwind.config.js
@@ -2,7 +2,7 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
-    "src/pages/**/*.{js,ts,jsx,tsx}",
+    "src/app/**/*.{js,ts,jsx,tsx}",
     "src/components/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
